% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-shape.R
\name{alpha_shape}
\alias{alpha_shape}
\title{alpha_shape}
\usage{
alpha_shape(simplicies, alpha_range, maxs, mins, n)
}
\arguments{
\item{simplicies}{A Delaunay trigulation list object created by 
\code{\link{delaunay}} or a alpha complex list object created by 
\code{\link{alpha_complex}}.}

\item{alpha_range, }{range of alpha value}

\item{mins}{Vector of length \code{n} listing the point space minimum for each dimension.
@param maxs Vector of length \code{n} listing the point space maximum for each  dimension.}

\item{n}{\code{n} dimension point co-ordinate}
}
\value{
grid stack as vector, gridSimplex, and the inputed grid point.
}
\description{
Compute an alpha Shape Grid using the Q-hull library.
}
\details{
The calculation is done by assigning the trigulation index when the grid cell center lies within the trigulation or -1 if it lies outside
}
\examples{
# Define points and create a Delaunay triangulation
x <- c(30, 70, 20, 50, 40, 70)
y <- c(35, 80, 70, 50, 60, 20)
p <- data.frame(x, y)
a_complex <- alpha_complex(points = p, alpha = 20)
alpha_shape(simplicies = a_complex,maxs = c(70,80),mins = c(20,20),n = 5,alpha_range = c(1:20))
}
