% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-shape.R
\docType{package}
\name{alphashape}
\alias{alphashape}
\title{Computation of n dimension \eqn{\alpha}-shape}
\description{
Implementation in n dimension of the alpha shape using the Q-hull library

\tabular{ll}{ Package: \tab alphashape\cr  Date: \tab
2019-03-14\cr License: \tab GPL-2\cr LazyLoad: \tab yes\cr }
}
\references{
http://www.qhull.org/html/qh-code.htm
}
\author{
Pascal Omondiagbe\cr Tom Etherington\cr Maintainers: pascal Omondiagbe <omondiagbep@landcareresearch.co.nz>
}
\keyword{package}
