% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex.R
\name{convex}
\alias{convex}
\title{Convex hull in N-dimensions.}
\usage{
convex(point = NULL, options = NULL)
}
\arguments{
\item{point}{\code{point} is an \code{n}-by-\code{dim} of dataframe or matrix. The rows of
\code{point} represent \code{n} points in \code{dim}-dimensional
space.}

\item{options}{String containing extra options for the underlying Qhull command.(See the Qhull documentation (\url{../doc/html/qdelaun.html}) for the available options.) The
\code{Qbb} option is always passed to Qhull. The default options are \code{Qt}.  The degenerate (zero area) regions are returned
For silent operation, specify the option \code{Pp}.}
}
\description{
This function calculates the  Convex hull in N-dimensions using the qhull library
}
\examples{
# define point
x = c(30,70,20,50,40,70)
y = c(35,80,70,50,60,20)
p =data.frame(x,y)
convex =convex(point=p)
}
\seealso{
Used internally by \code{\link{convex}}
}
