% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-simplex.R
\name{findSimplex}
\alias{findSimplex}
\title{Find Simplex}
\usage{
findSimplex(tri, inputPoint, testPoint)
}
\arguments{
\item{tri}{delaunay trigulation simplex using \code{\link{delaunay}}}

\item{inputPoint}{\code{n}-by-\code{dim} dataframe or matrix of original inputPoint.}

\item{testPoint}{\code{n}-by-\code{dim} dataframe of points to check.}
}
\value{
A \code{n*m} vector containing the result. -1  if a given point is outside the trigulation, or the trigulation index if the center of the cell is inside the alpha shape
}
\description{
Returns the simplicies of the delaunay trigulation which contains a given point.
}
\details{
Given a grid point and a test point point, the find Simplex will identify the simplicies contianing the test point. 
It works by first checking for all point inside a convex hull, and then check if the center of the grid cell is inside the trigulation
}
\examples{
 x = c(30,70,20,50,40,70)
 y = c(35,80,70,50,60,20)
 p = data.frame(x,y)
 v=voronoi(point =p)
 meshGrdiSpace = grid.coords(mins=c(15,0), maxs=c(35,200), nCoords=5)
 simplex <- findSimplex(v$tri, v$inputPoint,meshGrdiSpace)
}
