% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-generation.R
\name{gridGeneration}
\alias{gridGeneration}
\title{gridGeneration}
\usage{
gridGeneration(inputFile = NULL, point = NULL, nCoords,
  minPoint = NULL, maxPoint = NULL)
}
\arguments{
\item{nCoords}{\code{n} number of coordinates across the grid}

\item{minPoint}{Vector of length \code{n} listing the  minimum point for each dimension.}

\item{maxPoint}{Vector of length \code{n} listing the  maximum point for each dimension.}

\item{inputfile}{file name containing the input point and number of dimension}

\item{Point:}{Matrix. The input point.}
}
\value{
A \code{n*m} vector of matrix
 @export gridGeneration
}
\description{
Generate a \code{n*m} grid point by passing the vector of max and min point or file containing the initial data point or a matrix of input point
}
\details{
This function creates a grid by using the max/min values specify by the user. Note that when the user sepcify a matrix of point the maximum and minimum point will
          be derived from the input point and use when generating the grid (+ 0.3 is added to the max point, and -0.-3 is taken away from the min point). 
Point could be passed as a file or matrix of point
}
