% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalphashape.R
\name{nalphaShape}
\alias{nalphaShape}
\title{alphaShape}
\usage{
nalphaShape(point, alphaRange, maxs, mins, n)
}
\arguments{
\item{point}{observation as dataframe}

\item{alphaRange, }{range of alpha value}

\item{mins}{Vector of length \code{n} listing the point space minimum for each dimension.
@param maxs Vector of length \code{n} listing the point space maximum for each  dimension.}

\item{n}{\code{n} dimension point co-ordinate}
}
\value{
grid stack as vector, gridSimplex, and the inputted grid point.
}
\description{
Compute an alpha Shape Grid using the Q-hull library.
}
\details{
The calculation is done by assigning the trigulation index when the grid cell center lies within the trigulation or -1 if it lies outside
}
\examples{
 x = c(30,70,20,50,40,70,20)
 y = c(35,80,70,50,60,20,30)
 p = data.frame(x,y)
 nalphaShape(point = p,maxs = c(70,80),mins = c(20,20),n = 5,alphaRange = c(1:20))
}
