% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex-hull.R
\name{print.convex_hull}
\alias{print.convex_hull}
\title{Print Convex Hull Object}
\usage{
\method{print}{convex_hull}(x, ...)
}
\description{
A function the print the convex_hull class without the object.
}
\keyword{internal}
