% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{print.delaunay_tri}
\alias{print.delaunay_tri}
\title{Print Delaunay Trigulation Object}
\usage{
\method{print}{delaunay_tri}(x, ...)
}
\description{
A function the print the delaunay_tri class without the object.
}
\keyword{internal}
