% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{print.voronoi_diagram}
\alias{print.voronoi_diagram}
\title{Print Voronoi Diagram Object}
\usage{
\method{print}{voronoi_diagram}(x, ...)
}
\description{
A function the print the voronoi_diagram class without the object.
}
\keyword{internal}
