% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi diagram}
\usage{
voronoi(points = NULL, delaunay = FALSE)
}
\arguments{
\item{points}{a \eqn{n}-by-\eqn{d} dataframe or matrix. The rows
represent \eqn{n} points and the \eqn{d} columns the coordinates in 
\eqn{d}-dimensional space.}

\item{delaunay}{a boolean indicating if the Delaunay triangulation, which is 
the dual of the Voronoi diagram should also be returned, defaults to 
\code{FALSE}.}
}
\value{
Returns a list consisting of...  Additonally, if \code{delaunay = TRUE}...
}
\description{
This function calculates the 
\href{https://en.wikipedia.org/wiki/Voronoi_diagram}{Voronoi digram} of a set
of \eqn{n} points in \eqn{d}-dimensional space using the
\href{http://www.qhull.org}{Qhull} library.
}
\examples{
# Define points
x <- c(30, 70, 20, 50, 40, 70, 20, 55, 30)
y <- c(35, 80, 70, 50, 60, 20, 20, 55, 65)
p <- data.frame(x, y)
# Create Voronoi diagram and plot
vd <- voronoi(points = p)
cols = c("red", "blue", "green", "darkgrey", "purple", "lightseagreen",
         "brown", "darkgreen", "orange")
plot(vd$input_points, pch = as.character(seq(nrow(p))), col=cols,
     xlim=c(0,100), ylim=c(0,100))
r = 0
for (vd_region in vd$voronoi_regions) {
  r = r + 1
  if (!0 \%in\% vd_region) {
    polygon(vd$voronoi_vertices[vd_region,], density=20, col = cols[r])
  }
}

}
\references{
Barber CB, Dobkin DP, Huhdanpaa H (1996) The Quickhull algorithm 
for convex hulls. ACM Transactions on Mathematical Software, 22(4):469-83 
\url{https://doi.org/10.1145/235815.235821}.
}
\seealso{
\code{\link{delaunay}}
}
