% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex.R
\name{convex_hull}
\alias{convex_hull}
\title{convex_hull in d-dimensions.}
\usage{
convex_hull(points = NULL, options = NULL)
}
\arguments{
\item{points}{\code{points} is an \code{n}-by-\code{d} of dataframe or 
matrix. The rows of \code{points} represent \code{n} points in \code{d}-
dimensional space.}

\item{options}{String containing extra options for the underlying Qhull 
command.(See the Qhull documentation (\url{../doc/html/qdelaun.html}) for 
the available options.) The \code{Qbb} option is always passed to Qhull. 
The default options are \code{Qt}.  The degenerate (zero area) regions are
returned. For silent operation, specify the option \code{Pp}.}
}
\description{
This function calculates the convex hull in \code{d}-
dimensional space using the qhull library.
}
\examples{
# Define points
x = c(30,70,20,50,40,70)
y = c(35,80,70,50,60,20)
p = data.frame(x,y)
# Create convex hull and plot
ch = convex_hull(points = p)
plot(p, pch=as.character(seq(nrow(p))))
polygon(ch$setPoints, border="red")
}
\seealso{
Used internally by \code{\link{convex}}
}
