% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-simplex.R
\name{find_simplex}
\alias{find_simplex}
\title{Find simplex}
\usage{
find_simplex(simplicies, test_points)
}
\arguments{
\item{simplicies}{A Delaunay trigulation list object created by 
\code{\link{delaunay}} or a alpha complex list object created by 
\code{\link{alpha_complex} that contain simplicies}.}

\item{test_points}{a \eqn{n}-by-\eqn{d} dataframe or matrix. The rows
represent \eqn{n} points and the \eqn{d} columns the coordinates in 
\eqn{d}-dimensional space.}
}
\value{
A \eqn{n} length vector containing the index of the simplex the test 
point is within, or a value of NA if a test point is not within any of the 
simplicies.
}
\description{
Returns the simplicies of a Delaunay triangulation or alpha 
complex that contain the given set of test points.
}
\examples{
# Define points and create a Delaunay triangulation
x <- c(30, 70, 20, 50, 40, 70)
y <- c(35, 80, 70, 50, 60, 20)
p <- data.frame(x, y)
a_complex <- alpha_complex(points = p, alpha = 20)
# Check which simplex the test points belong to
p_test <- data.frame(c(20, 50, 60), c(20, 50, 60))
simplex <- find_simplex(simplicies = a_complex, test_points = p_test)
 
}
