% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-convex-hull.R
\name{in_convex_hull}
\alias{in_convex_hull}
\title{in convex hull}
\usage{
in_convex_hull(hull = NULL, test_points = NULL)
}
\arguments{
\item{points}{points to make convex hull}

\item{test_points:}{dataframe or matrix  \code{n}-by-\code{dim} of points to check.}
}
\value{
A \code{n*m} vector containing the result. True if a given point was inside the convexhull , otherwise false
}
\description{
To test if a given point is inside the convex hull. \code{TRUE}  if the point  lies within the hull and \code{FALSE} if it
lies outwith the hull
}
\examples{
# Define points to create the convex hull
x <- c(30, 70, 20, 50, 40, 70)
y <- c(35, 80, 70, 50, 60, 20)
p <- data.frame(x, y)
ch <- convex_hull(points = p)
# Check if some new points are in the convex hull
new = data.frame(c(20, 50), c(20, 50))
checks <- in_convex_hull(hull=ch, test_points = new)

}
